refactor: remove tsd usage
